Added Lost and Found Tab
NOTE: NOT thoroughly tested. Updating should not be a problem, but I have not tested what happens when there is no network connection. That should be done before an update.

Also: Need icon for lost and found
 table view cell in more tab implemented.

Added lost and found based on 2D Array on google site. requires manually adding items until firebase is integrated.